@@ -18,24 +18,26 @@
     "@fontsource/fredoka": "^5.2.10",
     "@fontsource/jetbrains-mono": "^5.2.8",
     "@fontsource/m-plus-rounded-1c": "^5.2.10",
-    "svelte": "^5.49.1"
+    "svelte": "^5.51.0"
   },
   "devDependencies": {
-    "@astrojs/check": "^0.9.7-beta.1",
-    "@astrojs/mdx": "^5.0.0-beta.2",
-    "@astrojs/svelte": "^8.0.0-beta.2",
+    "@astrojs/check": "0.9.7-beta.1",
+    "@astrojs/mdx": "5.0.0-beta.7",
+    "@astrojs/svelte": "8.0.0-beta.2",
+    "@eslint/js": "^9.39.2",
     "@tailwindcss/typography": "^0.5.19",
     "@tailwindcss/vite": "^4.1.18",
     "@types/hast": "^3.0.4",
     "@types/node": "^25.2.3",
-    "@typescript-eslint/eslint-plugin": "^8.54.0",
-    "@typescript-eslint/parser": "^8.54.0",
-    "astro": "^6.0.0-beta.5",
+    "@typescript-eslint/eslint-plugin": "^8.55.0",
+    "@typescript-eslint/parser": "^8.55.0",
+    "astro": "6.0.0-beta.11",
+    "astro-eslint-parser": "^1.3.0",
     "astro-og-canvas": "^0.10.1",
     "budoux": "^0.7.0",
     "canvaskit-wasm": "^0.40.0",
     "eslint": "^9.39.2",
-    "eslint-plugin-astro": "^1.5.0",
+    "eslint-plugin-astro": "^1.6.0",
     "hastscript": "^9.0.1",
     "postcss-html": "^1.8.1",
     "prettier": "^3.8.1",
@@ -43,7 +45,7 @@
     "rehype": "^13.0.2",
     "remark-denden-ruby": "^0.3.1",
     "remark-gemoji": "^8.0.0",
-    "stylelint": "^17.0.0",
+    "stylelint": "^17.3.0",
     "stylelint-config-html": "^1.1.0",
     "stylelint-config-recommended": "^18.0.0",
     "tailwindcss": "^4.1.18",
